added mips class

+++ memory.cpp
@@ -4,16 +4,16 @@
 
 namespace vCPU{
     Memory::Memory(){
-	    m_memspace = new uint8_t[MAX_ADDR];
+        m_memspace = new uint32_t[MAX_ADDR];
     }
     Memory::~Memory(){
         delete m_memspace;
     }
-    const uint8_t& Memory::read(uint8_t pos){
+    const uint32_t& Memory::read(uint8_t pos){
         return m_memspace[pos];
     }
 
-    void RAM::write(uint8_t pos, uint8_t value){
+    void RAM::write(uint8_t pos, uint32_t value){
         m_memspace[pos] = value;
     }
 }
